Use comment parser also in ImportExtractor
